src/requirements.txt
Flask
jsonschema
pyyaml